requirements.txt
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioitertools==0.12.0
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.25.1
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
black==23.11.0
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
camelot-py==0.11.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
deepl==1.15.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.40.0
fastapi==0.109.0
filelock==3.18.0
flower==2.0.1
fonttools==4.58.1
fpdf2==2.7.5
frozenlist==1.6.2
fsspec==2025.5.1
ghostscript==0.8.1
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-translate==3.11.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.62.3
h11==0.16.0
h2==3.2.0
hf-xet==1.1.3
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
humanize==4.12.3
hyperframe==5.2.0
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
kiwisolver==1.4.8
kombu==5.5.4
matplotlib==3.10.3
multidict==6.4.4
mypy_extensions==1.1.0
numpy==1.26.2
openai==1.23.2
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.1.3
pathspec==0.12.1
pdf2image==1.16.3
pdfminer.six==20250327
pdfplumber==0.11.6
pdftopng==0.2.4
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyMuPDF==1.26.0
pyparsing==3.2.3
pypdf==5.6.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.10
pytest==7.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
reportlab==4.0.7
requests==2.32.3
rfc3986==1.5.0
rsa==4.9.1
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
starlette==0.35.1
tabulate==0.9.0
tokenizers==0.21.1
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.27.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
